chore: fix up seo feature
