+++ src/main.tsx
@@ -7,7 +7,7 @@ import './i18n';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <App />
+    <App useRouter={true} />
   </React.StrictMode>,
 );
